refactor(streams): refactor streams.py into a package streams with the modules base_streams and signal_streams
